jenkins: move teardown code to postbuild task
As part of the teardown activities, we need to ensure
that the workspace is owned by the user that runs the
job. This commands were in the Build section, but if a
prior step fails, these commands are not executed, so
moving them to the Post-build task section, where they
will always run.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>