config,el8: Ensure there's no podman & buildah installed
Let's ensure there's no podman or buildah installed when we're
installing docker-ce, otherwise conflicts will happen due to the version
of the runc installed (and required) by podman & buildah.

This issue is only happening with the RHEL-8 image from Azure and podman
& buildah seem to be installed as part of the image, which explains why
we see this issue on RHEL-8 but not on CentOS 8. Regardless, I'm taking
the approach of calling `yum -y erase podman` on EL8 in general, as
those should be as close as possible to each other.

The error seen is:
```
Error:
 Problem: package podman-1.9.3-2.module+el8.2.1+6867+366c07d6.x86_64 requires runc >= 1.0.0-57, but none of the providers can be installed
  - package containerd.io-1.2.6-3.3.el7.x86_64 conflicts with containerd provided by containerd.io-1.3.7-3.1.el8.x86_64
  - package containerd.io-1.3.7-3.1.el8.x86_64 conflicts with runc provided by containerd.io-1.2.6-3.3.el7.x86_64
  - package containerd.io-1.3.7-3.1.el8.x86_64 conflicts with containerd provided by containerd.io-1.2.6-3.3.el7.x86_64
  - cannot install both containerd.io-1.3.7-3.1.el8.x86_64 and containerd.io-1.2.6-3.3.el7.x86_64
```

Fixes: #330

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>