jobs-builder: run SEV operator CI jobs on 'amd-ubuntu-2004_op-ci' nodes
The AMD SEV jobs for the operator CI should run on a different node than
those for Kata CI. This will make the operator CI SEV jobs to use the
'amd-ubuntu-2004_op-ci' labeled nodes (currently one in Jenkins pool).

The jobs/include/os2node.yaml.inc template uses the value of the project
name to decide which label to pick. It is fragile. I tried to use the
'template-name' property to be more accurate, however, the hifen in
'template-name' makes that property inaccessible from a jinja2 template...

Fixes #495
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>